requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
python-dotenv==1.0.0
gunicorn==20.1.0
argon2-cffi==21.3.0